support array class literal - only output known location type annotations

+++ src/share/classes/com/sun/tools/classfile/ExtendedAnnotation.java
@@ -64,7 +64,7 @@ public class ExtendedAnnotation {
     private static TypeAnnotations.Position read_position(ClassReader cr) throws IOException {
         // Copied from ClassReader
         TypeAnnotations.Position position = new TypeAnnotations.Position();
-        int tag = cr.readUnsignedByte();
+        int tag = (byte)cr.readUnsignedByte();  // cast to introduce signess
         TargetType type = TargetType.fromTargetTypeValue(tag);
 
         position.type = type;
